refactor(codex-adapter): extract load_history_inner into history_load.rs
Todo #247 task 3: session.rs was at ~900 lines; the loadHistory
temp-app-server + thread/read recursion doesn't need to live inline.
Verbatim move, de() promoted to pub(crate) so history_load can share it.
