+++ packages/core-rs/crates/mainframe-adapter-codex/src/lib.rs
@@ -24,6 +24,7 @@ pub mod external_sessions;
 pub mod history;
 pub(crate) mod history_collab;
 pub(crate) mod history_convert;
+pub(crate) mod history_load;
 pub(crate) mod image_generation_history;
 pub(crate) mod image_generation_render;
 pub mod item_types;
